pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dqs"
version = "0.1.0"
description = "Data Quality Scanner — parameterized SQL checks for cloud data warehouses"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pandas>=2.0",
    "sqlalchemy>=2.0",
    "duckdb>=0.10",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "numpy>=1.26",
    "faker>=24.0",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-connector-python>=3.0",
    "snowflake-sqlalchemy>=1.5",
]
bigquery = [
    "google-cloud-bigquery>=3.0",
    "sqlalchemy-bigquery>=1.9",
]
redshift = [
    "redshift-connector>=2.0",
]
databricks = [
    "databricks-sql-connector>=3.0",
]
postgres = [
    "psycopg2-binary>=2.9",
]
synapse = [
    "pyodbc>=4.0",
    "azure-identity>=1.15",
]
all = [
    "snowflake-connector-python>=3.0",
    "snowflake-sqlalchemy>=1.5",
    "google-cloud-bigquery>=3.0",
    "sqlalchemy-bigquery>=1.9",
    "redshift-connector>=2.0",
    "databricks-sql-connector>=3.0",
    "psycopg2-binary>=2.9",
    "pyodbc>=4.0",
    "azure-identity>=1.15",
]

[project.scripts]
dqs = "dqs.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["dqs"]
